Geom2D.CubicBezier.Linear

Description

Convenience wrapper around CubicBezier

Documentation

newtype AnyBezier a

A bezier curve of any degree.

Constructors

AnyBezier (Vector (V2 a))

data CubicBezier a

A cubic bezier curve.

Constructors

CubicBezier

Fields

cubicC0 :: !(V2 a)

cubicC1 :: !(V2 a)

cubicC2 :: !(V2 a)

cubicC3 :: !(V2 a)

data QuadBezier a

A quadratic bezier curve.

Constructors

QuadBezier

Fields

quadC0 :: !(V2 a)

quadC1 :: !(V2 a)

quadC2 :: !(V2 a)

data OpenPath a

Open cubicbezier path.

Constructors

OpenPath [(V2 a, PathJoin a)] (V2 a)

newtype ClosedPath a

Closed cubicbezier path.

Constructors

ClosedPath [(V2 a, PathJoin a)]

data PathJoin a

Join two points with either a straight line or a bezier curve with two control points.

Constructors

JoinLine

JoinCurve (V2 a) (V2 a)

newtype ClosedMetaPath a

Closed meta path.

Constructors

ClosedMetaPath [(V2 a, MetaJoin a)]

data OpenMetaPath a

Open meta path

Constructors

OpenMetaPath [(V2 a, MetaJoin a)] (V2 a)

data MetaJoin a

Join two meta points with either a bezier curve or tension contraints.

Constructors

MetaJoin

Fields

metaTypeL :: MetaNodeType a

tensionL :: Tension a

tensionR :: Tension a

metaTypeR :: MetaNodeType a

Controls (V2 a) (V2 a)

data MetaNodeType a

Node constraint type.

Constructors

Open

Curl

Fields

curlgamma :: a

Direction

Fields

nodedir :: V2 a

data FillRule

Describe the possible filling algorithms. Map the values of the `fill-rule` attributes.

Constructors

FillEvenOdd
Corresponds to the evenodd value.

FillNonZero
Corresponds to the nonzero value.

data Tension a

The tension value specifies how tense the curve is. A higher value means the curve approaches a line segment, while a lower value means the curve is more round. Metafont doesn't allow values below 3/4.

Constructors

Tension

Fields

tensionValue :: a

TensionAtLeast
Like Tension, but keep the segment inside the bounding triangle defined by the control points, if there is one.

Fields

tensionValue :: a

quadToCubic :: Fractional a => QuadBezier a -> CubicBezier a

Convert a quadratic bezier to a cubic bezier.

arcLength :: CubicBezier Double -> Double -> Double -> Double

arcLength c t tol finds the arclength of the bezier c at t, within given tolerance tol.

arcLengthParam :: CubicBezier Double -> Double -> Double -> Double

arcLengthParam c len tol finds the parameter where the curve c has the arclength len, within tolerance tol.

splitBezier :: (Unbox a, Fractional a, GenericBezier b) => b a -> a -> (b a, b a)

Split a bezier curve into two curves.

colinear :: CubicBezier Double -> Double -> Bool

Return False if some points fall outside a line with a thickness of the given tolerance.

evalBezier :: (GenericBezier b, Unbox a, Fractional a) => b a -> a -> V2 a

Calculate a value on the bezier curve.

evalBezierDeriv :: (Unbox a, Fractional a, GenericBezier b) => b a -> a -> (V2 a, V2 a)

Calculate a value and the first derivative on the curve.

bezierHoriz :: CubicBezier Double -> [Double]

Find the parameter where the bezier curve is horizontal.

bezierVert :: CubicBezier Double -> [Double]

Find the parameter where the bezier curve is vertical.

bezierSubsegment :: (Ord a, Unbox a, Fractional a, GenericBezier b) => b a -> a -> a -> b a

Return the subsegment between the two parameters.

reorient :: (GenericBezier b, Unbox a) => b a -> b a

Reorient to the curve B(1-t).

closedPathCurves :: Fractional a => ClosedPath a -> [CubicBezier a]

Return the closed path as a list of curves.

openPathCurves :: Fractional a => OpenPath a -> [CubicBezier a]

Return the open path as a list of curves.

curvesToClosed :: [CubicBezier a] -> ClosedPath a

Make an open path from a list of curves. The last control point of each curve is ignored.

closest :: CubicBezier Double -> V2 Double -> Double -> Double

Find the closest value on the bezier to the given point, within tolerance. Return the first value found.

unmetaOpen :: OpenMetaPath Double -> OpenPath Double

Create a normal path from a metapath.

unmetaClosed :: ClosedMetaPath Double -> ClosedPath Double

Create a normal path from a metapath.

union :: [ClosedPath Double] -> FillRule -> Double -> [ClosedPath Double]

`O((n+m)*log(n+m))`, for n segments and m intersections. Union of paths, removing overlap and rounding to the given tolerance.

bezierIntersection :: CubicBezier Double -> CubicBezier Double -> Double -> [(Double, Double)]

Find the intersections between two Bezier curves, using the Bezier Clip algorithm. Returns the parameters for both curves.

interpolateVector :: Num a => V2 a -> V2 a -> a -> V2 a

Interpolate between two vectors.

vectorDistance :: Floating a => V2 a -> V2 a -> a

Distance between two vectors.

findBezierInflection :: CubicBezier Double -> [Double]

Find inflection points on the curve.

findBezierCusp :: CubicBezier Double -> [Double]

Find the cusps of a bezier.