Reanimate.Constants

Description

Reanimate configures a consistent, default canvas. The values of this default can be observed via the constants in this module. Keep in mind, these values describe the default canvas and will not apply to custom viewports.

Documentation

screenWidth :: Fractional a => a

Number of units from the left-most point to the right-most point on the screen.

screenHeight :: Fractional a => a

Number of units from the bottom to the top of the screen.

screenTop :: Fractional a => a

Position of the top of the screen.

screenBottom :: Fractional a => a

Position of the bottom of the screen.

screenLeft :: Fractional a => a

Position of the left side of the screen.

screenRight :: Fractional a => a

Position of the right side of the screen.

defaultDPI :: Dpi

SVG allows measurements in inches which have to be converted to local units. This value describes how many local units there are in an inch.

defaultStrokeWidth :: Double

Default thickness of lines.