Reanimate.Math.Common

Description

Low-level primitives related to computational geometry.

Ring

newtype Ring a

Circular collection of pairs.

Constructors

Ring (Vector (V2 a))

ringSize :: Ring a -> Int

Number of elements in the ring.

ringAccess :: Ring a -> Int -> V2 a

Safe method for accessing elements in the ring.

ringClamp :: Ring a -> Int -> Int

Clamp index to within the usable range for the ring.

ringUnpack :: Ring a -> Vector (V2 a)

Convert ring to a vector.

ringPack :: Vector (V2 a) -> Ring a

Convert vector to a ring.

ringMap :: (V2 a -> V2 b) -> Ring a -> Ring b

Map each element of a ring.

ringRayIntersect :: Ring Rational -> (Int, Int) -> (Int, Int) -> Maybe (V2 Rational)

Compute the intersection of two pairs of nodes in the ring.

Math

area :: Fractional a => V2 a -> V2 a -> V2 a -> a

Compute area of triangle.

area2X :: Fractional a => V2 a -> V2 a -> V2 a -> a

Compute 2x area of triangle. This avoids a division.

isLeftTurn :: (Fractional a, Ord a, Epsilon a) => V2 a -> V2 a -> V2 a -> Bool

Return True iff the line from p1 to p2 makes a left-turn to p3.

isLeftTurnOrLinear :: (Fractional a, Ord a, Epsilon a) => V2 a -> V2 a -> V2 a -> Bool

Return True iff the line from p1 to p2 does not make a right-turn to p3.

isRightTurn :: (Fractional a, Ord a, Epsilon a) => V2 a -> V2 a -> V2 a -> Bool

Return True iff the line from p1 to p2 makes a right-turn to p3.

isRightTurnOrLinear :: (Fractional a, Ord a, Epsilon a) => V2 a -> V2 a -> V2 a -> Bool

Return True iff the line from p1 to p2 does not make a left-turn to p3.

direction :: Num a => V2 a -> V2 a -> V2 a -> a

Compute the change in direction in a line between the three points.

isInside :: (Fractional a, Ord a) => V2 a -> V2 a -> V2 a -> V2 a -> Bool

Returns True if the fourth argument is inside the triangle or on the border.

isInsideStrict :: (Fractional a, Ord a) => V2 a -> V2 a -> V2 a -> V2 a -> Bool

Returns True iff the fourth argument is inside the triangle.

barycentricCoords :: Fractional a => V2 a -> V2 a -> V2 a -> V2 a -> (a, a, a)

Compute relative coordinates inside the triangle. Invariant: a+b+c=1

rayIntersect :: (Fractional a, Ord a) => (V2 a, V2 a) -> (V2 a, V2 a) -> Maybe (V2 a)

Compute intersection of two infinite lines.

isBetween :: (Ord a, Fractional a) => V2 a -> (V2 a, V2 a) -> Bool

Returns True iff a point is on a line segment.

lineIntersect :: (Ord a, Fractional a) => (V2 a, V2 a) -> (V2 a, V2 a) -> Maybe (V2 a)

Compute intersection of two line segments.

distSquared :: Num a => V2 a -> V2 a -> a

Compute the square of the distance between two points.

approxDist :: (Real a, Fractional a) => V2 a -> V2 a -> a

Approximate the distance between two points.

distance' :: (Real a, Fractional a) => V2 a -> V2 a -> Double

Approximate the distance between two points.

triangleAngles :: V2 Double -> V2 Double -> V2 Double -> (Double, Double, Double)

Approximate the angles of a triangle.

class Num a => Epsilon a where

Provides a fairly subjective test to see if a quantity is near zero.

>>> nearZero (1e-11 :: Double)
False

>>> nearZero (1e-17 :: Double)
True

>>> nearZero (1e-5 :: Float)
False

>>> nearZero (1e-7 :: Float)
True

Methods

nearZero :: a -> Bool

Determine if a quantity is near zero.

Orphan instances

Epsilon Rational

Instance details

Methods

nearZero :: Rational -> Bool