Reanimate.Morph.Common

Description

Documentation

type PointCorrespondence = Polygon -> Polygon -> (Polygon, Polygon)

Method determining how points in the source polygon align with points in the target polygon.

type Trajectory = (Polygon, Polygon) -> Double -> Polygon

Method for interpolating between two aligned polygons.

type ObjectCorrespondence = [GPolygon] -> [GPolygon] -> [(GPolygon, GPolygon)]

Method for pairing sets of polygons.

data Morph

Morphing strategy

Constructors

Morph

Fields

morphTolerance :: Double

Morphing curves is not always possible and sometimes shapes are reduced to polygons or meta-curves. This parameter determined the accuracy of this transformation.

morphColorComponents :: ColorComponents

Color components used for color interpolation. LAB is usually the best option here.

morphPointCorrespondence :: PointCorrespondence

Desired point-correspondence algorithm.

morphTrajectory :: Trajectory

Desired interpolation algorithm.

morphObjectCorrespondence :: ObjectCorrespondence

Desired object-correspondence algorithm.

morph :: Morph -> SVG -> SVG -> Double -> SVG

Apply morphing strategy to interpolate between two SVG images.

splitObjectCorrespondence :: ObjectCorrespondence

Object-correspondence algorithm that splits objects in smaller pieces as necessary.

dupObjectCorrespondence :: ObjectCorrespondence

Object-correspondence algorithm that duplicate objects as necessary.

genesisObjectCorrespondence :: ObjectCorrespondence

Object-correspondence algorithm that spawn objects as necessary.

toShapes :: Double -> SVG -> [(DrawAttributes, Polygon)]

Extract shapes and their graphical attributes from an SVG node.

normalizePolygons :: (Real a, Fractional a, Epsilon a) => APolygon a -> APolygon a -> (APolygon a, APolygon a)

Add points to each polygon such that they end up with same size.

annotatePolygons :: (Polygon -> SVG) -> SVG -> SVG

Map over each polygon in an SVG node.

unsafeSVGToPolygon :: Double -> SVG -> Polygon

Extract the first polygon in an SVG node. Will fail if there are no acceptable shapes.