Reanimate.Parameters

Description

Parameters define the global context of an animation. They are set once before an animation is rendered and may not change during rendering.

Documentation

data Raster

Raster engines turn SVG images into pixels.

Constructors

RasterNone
Do not use any external raster engine. Rely on the browser or ffmpeg.

RasterAuto
Scan for installed raster engines and pick the fastest one.

RasterInkscape
Use Inkscape to raster SVG images.

RasterRSvg
Use rsvg-convert to raster SVG images.

RasterMagick
Use imagemagick to raster SVG images.

type Width = Int

Width of animation in pixels.

type Height = Int

Height of animation in pixels.

type FPS = Int

Framerate of animation in frames per second.

pRaster :: Raster

Selected raster engine.

pFPS :: FPS

Selected framerate.

pWidth :: Width

Width of animation in pixel.

pHeight :: Height

Height of animation in pixel.

pNoExternals :: Bool

This parameter determined whether or not external tools are allowed. If this flag is True then tools such as latex and blender will not be invoked.

pRootDirectory :: FilePath

Root directory of animation. Images and other data has to be placed here if they are referenced in an SVG image.

setRaster :: Raster -> IO ()

Set raster engine.

setFPS :: FPS -> IO ()

Set desired framerate.

setWidth :: Width -> IO ()

Set desired width of animation in pixel.

setHeight :: Height -> IO ()

Set desired height of animation in pixel.

setNoExternals :: Bool -> IO ()

Set whether external tools are allowed.

setRootDirectory :: FilePath -> IO ()

Set the root animation directory.