Reanimate.PolyShape

Description

A PolyShape is a closed set of curves.

Documentation

newtype PolyShape

Shape drawn by continuous line. May have overlap, may be convex.

Constructors

PolyShape

Fields

unPolyShape :: ClosedPath Double

data PolyShapeWithHoles

Polyshape with smaller, fully-enclosed holes.

svgToPolyShapes :: Tree -> [PolyShape]

Extract all shapes from SVG nodes. Drawing attributes such as stroke and fill color are discarded.

svgToPolygons :: Double -> SVG -> [Polygon]

Extract all polygons from SVG nodes. Curves are approximated to within the given tolerance.

renderPolyShape :: PolyShape -> Tree

Render a polyshape as a single SVG path.

renderPolyShapes :: [PolyShape] -> Tree

Render a set of polyshapes as a single SVG path.

renderPolyShapePoints :: PolyShape -> Tree

Render control-points of a polyshape as circles.

plPathCommands :: PolyShape -> [PathCommand]

Convert a polyshape to a list of SVG path commands.

plLineCommands :: PolyShape -> [LineCommand]

Convert a polyshape to a list of line commands.

plLength :: PolyShape -> Double

Length of polyshape circumference.

plArea :: PolyShape -> Double

Area of polyshape.

plCurves :: PolyShape -> [CubicBezier Double]

Destruct a polyshape into constituent curves.

isInsideOf :: PolyShape -> PolyShape -> Bool

True iff lhs is inside of rhs. lhs and rhs may not overlap. Implementation: Trace a vertical line through the origin of A and check of this line intersects and odd number of times on both sides of A.

plFromPolygon :: [RPoint] -> PolyShape

Construct a polyshape from the vertices in a polygon.

plToPolygon :: Double -> PolyShape -> Polygon

Approximate a polyshape as a polygon within the given tolerance.

plDecompose :: [PolyShape] -> [[RPoint]]

Deconstruct a polyshape into non-intersecting, convex polygons.

unionPolyShapes :: [PolyShape] -> [PolyShape]

Merge overlapping shapes.

unionPolyShapes' :: Double -> [PolyShape] -> [PolyShape]

Merge overlapping shapes to within given tolerance.

plDecompose' :: Double -> [PolyShape] -> [[RPoint]]

Deconstruct a polyshape into non-intersecting, convex polygons.

decomposePolygon :: [RPoint] -> [[RPoint]]

Split polygon into smaller, convex polygons.

plGroupShapes :: [PolyShape] -> [PolyShapeWithHoles]

Find holes and group them with their parent.

mergePolyShapeHoles :: PolyShapeWithHoles -> PolyShape

Cut out holes.

plPartial :: Double -> PolyShape -> PolyShape

Partially draw polyshape.

plGroupTouching :: [PolyShape] -> [[([RPoint], PolyShape)]]

Find intersection points.