Reanimate.Povray

Description

Povray is a scriptable raytracer. All povray functions are cached and will reuse images when scripts stay the same.

Documentation

povray :: [String] -> Text -> Tree

Run the povray raytracer with a default resolution of 320x180 and antialiasing enabled. The resulting image is scaled to fit the screen exactly.

povrayQuick :: [String] -> Text -> Tree

Run the povray raytracer with a default resolution of 320x180 but without antialiasing. The resulting image is scaled to fit the screen exactly.

povraySlow :: [String] -> Text -> Tree

Run the povray raytracer with a default resolution of 1440x2560 and antialiasing enabled. The FilePath points to a PNG file containing the resulting image.

povrayExtreme :: [String] -> Text -> Tree

Run the povray raytracer with a default resolution of 2160x3840 and antialiasing enabled. The FilePath points to a PNG file containing the resulting image.

povray' :: [String] -> Text -> FilePath

Run the povray raytracer with a default resolution of 320x180 and antialiasing enabled. The FilePath points to a PNG file containing the resulting image.

povrayQuick' :: [String] -> Text -> FilePath

Run the povray raytracer with a default resolution of 320x180 but without antialiasing. The FilePath points to a PNG file containing the resulting image.

povraySlow' :: [String] -> Text -> FilePath

Run the povray raytracer with a default resolution of 1440x2560 and antialiasing enabled. The FilePath points to a PNG file containing the resulting image.

povrayExtreme' :: [String] -> Text -> FilePath

Run the povray raytracer with a default resolution of 2160x3840 and antialiasing enabled. The FilePath points to a PNG file containing the resulting image.