Reanimate.Render

Description

Internal tools for rastering SVGs and rendering videos. You are unlikely to ever directly use the functions in this module.

Documentation

render :: Animation -> FilePath -> Raster -> Format -> Width -> Height -> FPS -> Bool -> IO ()

Render animation to a video file with given parameters.

renderSvgs :: FilePath -> Int -> Bool -> Animation -> IO ()

Generate SVGs at 60fps and put them in a folder.

renderSvgs_ :: Animation -> (Int -> FilePath -> IO ()) -> IO ()

renderSnippets :: Animation -> IO ()

Render 10 frames and print them to stdout. Used for testing.

XXX: Not related to the snippets in the playground.

renderLimitedFrames :: FilePath -> Int -> Bool -> Int -> Animation -> IO ()

Render as many frames as possible in 2 seconds. Limited to 20 frames.

data Format

Video formats supported by reanimate.

Constructors

RenderMp4

RenderGif

RenderWebm

data Raster

Raster engines turn SVG images into pixels.

Constructors

RasterNone
Do not use any external raster engine. Rely on the browser or ffmpeg.

RasterAuto
Scan for installed raster engines and pick the fastest one.

RasterInkscape
Use Inkscape to raster SVG images.

RasterRSvg
Use rsvg-convert to raster SVG images.

RasterMagick
Use imagemagick to raster SVG images.

type Width = Int

Width of animation in pixels.

type Height = Int

Height of animation in pixels.

type FPS = Int

Framerate of animation in frames per second.

requireRaster :: Raster -> IO Raster

Resolve RasterNone and RasterAuto. If no valid raster can be found, exit with an error message.

selectRaster :: Raster -> IO Raster

Resolve RasterNone and RasterAuto. If no valid raster can be found, return RasterNone.

applyRaster :: Raster -> FilePath -> IO ()

Convert SVG file to a PNG file with selected raster engine. If raster engine is RasterAuto or RasterNone, do nothing.