Reanimate.Svg.Unuse

Description

Documentation

replaceUses :: Document -> Document

Replace all use nodes with their definition.

unbox :: Document -> Tree

Transform out viewbox. Definitions and CSS rules are discarded.

unboxFit :: Document -> Tree

Transform out viewbox and fit image to screen size.

embedDocument :: Document -> Tree

Embed Document. This keeps the entire document intact but makes it more difficult to use, say, pathify on it.