Reanimate.Voice

Description

Reanimate can automatically synchronize animations to your voice if you have a transcript and an audio recording. This works with the help of Gentle (https://lowerquality.com/gentle/). Accuracy is not perfect but it is pretty close, and it is by far the easiest way of adding narration to an animation.

Documentation

data Transcript

Aligned transcript. Contains the transcript text as well as timing data for each word.

Constructors

Transcript

Fields

transcriptText :: Text

transcriptKeys :: Map Text Int

transcriptWords :: [TWord]

data TWord

Spoken word. Includes information about when it was spoken, its duration, and its phonemes.

Constructors

TWord

Fields

wordAligned :: Text

wordCase :: Text

wordStart :: Double

Start of pronunciation in seconds

wordStartOffset :: Int

Character index of word in transcript

wordEnd :: Double

End of pronunciation in seconds

wordEndOffset :: Int

Last character index of word in transcript

wordPhones :: [Phone]

wordReference :: Text

The word being pronounced.

data Phone

Phoneme type

Constructors

Phone

Fields

phoneDuration :: Double

phoneType :: Text

findWord :: Transcript -> [Text] -> Text -> TWord

Locate the first word that occurs after all the given keys. An error is thrown if no such word exists. An error is thrown if the keys do not exist in the transcript.

findWords :: Transcript -> [Text] -> Text -> [TWord]

Locate all words that occur after all the given keys. May return an empty list. An error is thrown if the keys do not exist in the transcript.

loadTranscript :: FilePath -> Transcript

Loading a transcript does three things depending on which files are available with the same basename as the input argument: 1. If a JSON file is available, it is parsed and returned. 2. If an audio file is available, reanimate tries to align it by calling out to Gentle on localhost:8765/. If Gentle is not running, an error will be thrown. 3. If only the text transcript is available, a fake transcript is returned, with timings roughly at 120 words per minute.

fakeTranscript :: Text -> Transcript

Fake transcript timings at roughly 120 words per minute.

splitTranscript :: Transcript -> [(SVG, TWord)]

Convert the transcript text to an SVG image using LaTeX and associate each word image with its timing information.

annotateWithTranscript :: Transcript -> Scene s ()

Helper function for rendering a transcript.